Split record and collection components
