+++ packages/core/src/recordDS.js
@@ -361,4 +361,9 @@ function defineRequestMethods(recordDS, privateData) {
   Object.defineProperty(recordDS, 'create', { value: create })
   Object.defineProperty(recordDS, 'update', { value: update })
   Object.defineProperty(recordDS, 'delete', { value: deleteFn })
+  Object.defineProperty(recordDS, 'refetch', {
+    value: function () {
+      return recordDS.fetch(privateData.id)
+    }
+  })
 }
