+++ packages/ui/src/LoadingIndicator/__tests__/LoadingIndicator.test.js
@@ -1,31 +1,33 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import { Spinner } from '@blueprintjs/core'
-import DataSourceComponent from '../DataSourceComponent'
-import LoadingIndicator from '../LoadingIndicator'
+import CollectionComponent from '../../CollectionComponent'
+import LoadingIndicator from '../index'
 
 describe('<LoadingIndicator />', () => {
   test('busy dataSource', () => {
     const dataSource = {
+      __isCollectionDS__: true,
       status: 'busy'
     }
 
     const comp = shallow(<LoadingIndicator dataSource={dataSource} />)
     const spinner = comp.renderProp('render')()
 
-    expect(comp.is(DataSourceComponent)).toBe(true)
+    expect(comp.is(CollectionComponent)).toBe(true)
     expect(spinner.is(Spinner)).toBe(true)
     expect(spinner.prop('size')).toBe(16)
   })
 
   test('ready dataSource', () => {
     const dataSource = {
+      __isCollectionDS__: true,
       status: 'ready'
     }
     const comp = shallow(<LoadingIndicator dataSource={dataSource} />)
     const spinner = comp.renderProp('render')()
 
-    expect(comp.is(DataSourceComponent)).toBe(true)
+    expect(comp.is(CollectionComponent)).toBe(true)
     expect(spinner.type()).toBeNull()
   })
 })
