+++ packages/ui/src/Toolbar/__tests__/Toolbar.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import Toolbar from '../Toolbar'
+import Toolbar from '../index'
 
 describe('<Toolbar />', () => {
   test('default toolbar', () => {
