+++ packages/ui/src/index.js
@@ -1,17 +1,19 @@
-import DataSourceComponent from './DataSourceComponent'
+import CollectionComponent from './CollectionComponent'
 import Error from './Error'
 import LoadingIndicator from './LoadingIndicator'
 import Pagination from './Pagination'
+import RecordComponent from './RecordComponent'
 import RefreshAction from './RefreshAction'
 import Table from './Table'
 import Toolbar from './Toolbar'
 import View from './View'
 
 export {
-  DataSourceComponent,
+  CollectionComponent,
   Error,
   LoadingIndicator,
   Pagination,
+  RecordComponent,
   RefreshAction,
   Table,
   Toolbar,
